refactor: extract discovery.py + card_resources.py from __init__.py
Splits two pieces of __init__.py into their own modules. __init__.py
shrinks from 1032 to 744 LOC; each new module is self-contained and
imports DOMAIN/coordinators directly rather than reaching back into
the package root.

- discovery.py — _discover_cameras, _discover_locks,
  _schedule_lock_config_retry, _async_discover_devices, plus the
  _LOCK_CONFIG_RETRY_DELAYS backoff schedule. Background discovery
  is best-effort; failures are logged not raised.
- card_resources.py — _register_card_resource and
  _unregister_card_resource for the three Lovelace cards, with the
  add_extra_js_url fallback path.
- __init__.py re-exports both modules' public symbols
  (_discover_cameras, _async_discover_devices, _register_card_resource,
  etc.) so existing tests that import them via the package root keep
  working unchanged. The frontend module re-export is kept for the
  same reason — tests patch custom_components.verisure_owa.frontend.

Note: the larger decomposition items (models.py / responses.py split,
alarm_control_panel.py split, client.py split) are not in this PR.
Each one is a substantial structural change with its own integration
risk and warrants its own focused PR.
